+++ testing/io/dicom/DicomFileReaderHelperTest.cpp
@@ -77,7 +77,7 @@ namespace rttb
 			//2) test getFileNames() with a RTDOSE file name and check equal with getFileNamesWithSameUID()
 			std::vector<FileNameType> fileVector2 = rttb::io::dicom::getFileNames(RTDOSE_FILENAME);
 
-			for (int i = 0; i < fileVector.size(); i++)
+			for (size_t i = 0; i < fileVector.size(); i++)
 			{
 				CHECK_EQUAL(fileVector.at(i), fileVector2.at(i));
 			}
